+++ foliant/preprocessors/pgsqldoc/pgsqldoc.py
@@ -3,19 +3,86 @@ Preprocessor for Foliant documentation authoring tool.
 Generates documentation from PostgreSQL database structure,
 '''
 
-
-import re
-import os
 import traceback
 import psycopg2
-import yaml
-from shutil import copyfile
 from jinja2 import Environment, FileSystemLoader
 from pkg_resources import resource_filename
 from foliant.preprocessors.base import BasePreprocessor
 from .queries import (TablesQuery, ColumnsQuery, ForeignKeysQuery,
-                      FunctionsQuery, ParametersQuery, TriggersQuery,
-                      SCHEMA, TABLE_NAME)
+                      FunctionsQuery, ParametersQuery, TriggersQuery)
+from .utils import copy_if_not_exists
+from .combined_options import CombinedOptions, yaml_to_dict_conversion
+from foliant.utils import output
+from copy import deepcopy
+
+
+def collect_datasets(connection,
+                     filters: dict) -> dict:
+
+        result = {}
+
+        tables = TablesQuery(connection, filters).run()
+        columns = ColumnsQuery(connection, filters).run()
+        fks = ForeignKeysQuery(connection, filters).run()
+
+        # fill each table with columns and foreign keys
+        result['tables'] = collect_tables(tables, columns, fks)
+
+        functions = FunctionsQuery(connection, filters).run()
+        parameters = ParametersQuery(connection, filters).run()
+
+        # fill each function with its parameters
+
+        result['functions'] = collect_functions(functions, parameters)
+
+        result['triggers'] = TriggersQuery(connection, filters).run()
+        return result
+
+
+def collect_tables(tables: list,
+                   columns: list,
+                   fks: list) -> list:
+    '''
+    Parse table and column query results got from db and:
+
+    - add 'columns' attribute to each table row with list of table columns;
+    - add 'foreign_keys' attribute to each column with list of fks if it is a
+      forign key column.
+
+    returns transformed list of tables.
+    '''
+
+    result = deepcopy(tables)
+    for table in result:
+        # get columns for this table
+        table_columns = list(filter(lambda x: x['table_name'] == table['relname'],
+                                    columns))
+        for col in table_columns:
+            # get foreign keys for this column
+            fks_fltr = list(filter(lambda x:
+                                   x['table_name'] == table['relname'] and
+                                   x['column_name'] == col['column_name'],
+                                   fks))
+            col['foreign_keys'] = fks_fltr
+        table['columns'] = table_columns
+    return result
+
+
+def collect_functions(functions: list,
+                      parameters: list) -> list:
+    '''
+    Parse function and parameter query results got from db and add 'parameters'
+    key to each function filled with its parameters
+    '''
+
+    result = deepcopy(functions)
+    for func in result:
+        # get parameters for this function
+        function_params = list(filter(lambda x: x['specific_name'] ==
+                                      func['specific_name'],
+                                      parameters))
+        func['parameters'] = function_params
+    return result
 
 
 class Preprocessor(BasePreprocessor):
@@ -43,47 +110,6 @@ class Preprocessor(BasePreprocessor):
         self._env = \
             Environment(loader=FileSystemLoader(str(self.project_path)))
 
-    def _collect_datasets(self,
-                          filters: dict,
-                          draw: bool) -> dict:
-
-        result = {}
-
-        tables = TablesQuery(self._con, filters).run()
-        columns = ColumnsQuery(self._con, filters).run()
-        fks = ForeignKeysQuery(self._con, filters).run()
-
-        # fill each table with columns and foreign keys
-        for table in tables:
-            columns_fltr = list(filter(lambda x:
-                                       x['table_name'] == table['relname'],
-                                       columns))
-            for col in columns_fltr:
-                fks_fltr = list(filter(lambda x:
-                                       x['table_name'] == table['relname'] and
-                                       x['column_name'] == col['column_name'],
-                                       fks))
-                col['foreign_keys'] = fks_fltr
-            table['columns'] = columns_fltr
-
-        result['tables'] = tables
-
-        functions = FunctionsQuery(self._con, filters).run()
-        parameters = ParametersQuery(self._con, filters).run()
-
-        # fill each function with its parameters
-        for func in functions:
-            params_fltr = list(filter(lambda x: x['specific_name'] ==
-                                      func['specific_name'],
-                                      parameters))
-            func['parameters'] = params_fltr
-
-        result['functions'] = functions
-
-        triggers = TriggersQuery(self._con, filters).run()
-        result['triggers'] = triggers
-        return result
-
     def _to_md(self,
                data: dict,
                doc_template: str) -> str:
@@ -93,7 +119,7 @@ class Preprocessor(BasePreprocessor):
                                      functions=data['functions'],
                                      triggers=data['triggers'])
         except Exception as e:
-            print(f'\nFailed to render doc template {doc_template}:', e)
+            output(f'\nFailed to render doc template {doc_template}:', self.quiet)
             info = traceback.format_exc()
             self.logger.debug(f'Failed to render doc template:\n\n{info}')
             return ''
@@ -112,65 +138,74 @@ class Preprocessor(BasePreprocessor):
         return result
 
     def _gen_docs(self,
-                  filters: dict,
-                  draw: bool,
-                  doc_template: str,
-                  scheme_template: str) -> str:
-        data = self._collect_datasets(filters, draw)
-        docs = self._to_md(data, doc_template)
-        if draw:
-            docs += '\n\n' + self._to_diag(data, scheme_template)
+                  options: CombinedOptions) -> str:
+        data = collect_datasets(self._con, options['filters'])
+        docs = self._to_md(data, options['doc_template'])
+        if options['draw']:
+            docs += '\n\n' + self._to_diag(data,
+                                           options['scheme_template'])
         return docs
 
-    def process_pgsqldoc_blocks(self, content: str) -> str:
-        def _sub(block: str) -> str:
-            if block.group('options'):
-                tag_options = self.get_options(block.group('options'))
-            else:
-                tag_options = {}
-            host = tag_options.get('host', self.options['host'])
-            port = tag_options.get('port', self.options['port'])
-            dbname = tag_options.get('dbname', self.options['dbname'])
-            user = tag_options.get('user', self.options['user'])
-            password = tag_options.get('password', self.options['password'])
+    def _connect(self, options: CombinedOptions):
+        """
+        Connect to PostgreSQL database using parameters from options.
+        Save connection object into self._con.
+
+        options(CombinedOptions) — CombinedOptions object with options from tag
+                                   and config.
+        """
+        try:
             self._con = None
-            try:
-                self.logger.debug(f'Trying to connect: host={host} port={port}'
-                                  f' dbname={dbname}, user={user} '
-                                  f'password={password}.')
-                self._con = psycopg2.connect(f"host='{host}' "
-                                             f"port='{port}' "
-                                             f"dbname='{dbname}' "
-                                             f"user='{user}'"
-                                             f"password='{password}'")
-            except psycopg2.OperationalError as e:
-                info = traceback.format_exc()
-                print(f'\nFailed to connect to host {host}. '
-                      'Documentation was not generated')
-                self.logger.debug(f'Failed to connect: host={host} port={port}'
-                                  f' dbname={dbname}, user={user} '
-                                  f'password={password}.\n\n{info}')
+            self.logger.debug(f"Trying to connect: host={options['host']} port={options['port']}"
+                              f" dbname={options['dbname']}, user={options['user']} "
+                              f"password={options['password']}.")
+            self._con = psycopg2.connect(f"host='{options['host']}' "
+                                         f"port='{options['port']}' "
+                                         f"dbname='{options['dbname']}' "
+                                         f"user='{options['user']}'"
+                                         f"password='{options['password']}'")
+        except psycopg2.OperationalError:
+            info = traceback.format_exc()
+            output(f"\nFailed to connect to host {options['host']}. "
+                   'Documentation was not generated', self.quiet)
+            self.logger.debug(f'Failed to connect: host={options["host"]}'
+                              f' port={options["port"]}'
+                              f' dbname={options["dbname"]} '
+                              f'user={options["user"]} '
+                              f'password={options["password"]}.\n\n{info}')
+            raise psycopg2.OperationalError
+
+    def _create_default_templates(self, options: CombinedOptions):
+        """
+        Copy default templates to project dir if their names in options are
+        same as default.
+        """
+
+        if options.is_default('doc_template'):
+            source = self.project_path / options['doc_template']
+            to_copy = resource_filename(__name__,
+                                        f"templates/{options.defaults['doc_template']}")
+            copy_if_not_exists(source, to_copy)
+
+        if options.is_default('scheme_template'):
+            source = self.project_path / options['scheme_template']
+            to_copy = resource_filename(__name__,
+                                        f"templates/{options.defaults['scheme_template']}")
+            copy_if_not_exists(source, to_copy)
+
+    def process_pgsqldoc_blocks(self, content: str) -> str:
+        def _sub(block) -> str:
+            tag_options = self.get_options(block.group('options'))
+            options = CombinedOptions(self.options,
+                                      tag_options,
+                                      conversions={'filters': yaml_to_dict_conversion},
+                                      defaults=self.defaults)
+            self._connect(options)
+            if not self._con:
                 return ''
-            if 'filters' in tag_options:
-                filters = yaml.load(tag_options['filters'])
-            else:
-                filters = self.options['filters']
 
-            draw = tag_options.get('draw', self.options['draw'])
-
-            doc_template = tag_options.get('doc_template',
-                                           self.options['doc_template'])
-            scheme_template = tag_options.get('scheme_template',
-                                              self.options['scheme_template'])
-            if doc_template == self.defaults['doc_template'] and\
-                    not os.path.exists(self.project_path / doc_template):
-                copyfile(resource_filename(__name__, 'templates/pgsqldoc.j2'),
-                         self.project_path / doc_template)
-            if scheme_template == self.defaults['scheme_template'] and\
-                    not os.path.exists(self.project_path / scheme_template):
-                copyfile(resource_filename(__name__, 'templates/scheme.j2'),
-                         self.project_path / scheme_template)
-            return self._gen_docs(filters, draw, doc_template, scheme_template)
+            self._create_default_templates(options)
+            return self._gen_docs(options)
         return self.pattern.sub(_sub, content)
 
     def apply(self):
@@ -182,7 +217,9 @@ class Preprocessor(BasePreprocessor):
             with open(markdown_file_path, encoding='utf8') as markdown_file:
                 content = markdown_file.read()
 
+            processed_content = self.process_pgsqldoc_blocks(content)
+
             with open(markdown_file_path, 'w', encoding='utf8') as markdown_file:
-                markdown_file.write(self.process_pgsqldoc_blocks(content))
+                markdown_file.write(processed_content)
 
         self.logger.info('Preprocessor applied')