+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.1.3
+
+-    Add tests; refactor code
+-    Fix triggers and functions; add description to functions
+-    Fix template
+
 # 1.1.2
 
 -    Queries are now ordered (not adjustable right now)